refactor: stake, dealerStatus 책임 수정
